remove extra themed files
