+++ .babelrc
@@ -16,6 +16,6 @@
       "compact": false
     }
   ],
-  "plugins": ["macros", "module-resolver", "babel-plugin-styled-components"],
+  "plugins": ["module-resolver", "babel-plugin-styled-components"],
   "presets": ["@babel/env", "module:metro-react-native-babel-preset"]
 }
