+++ package.json
@@ -118,8 +118,6 @@
     "@babel/plugin-proposal-object-rest-spread": "^7.11.0",
     "@babel/polyfill": "^7.11.5",
     "@babel/preset-env": "^7.11.5",
-    "@babel/preset-flow": "^7.10.4",
-    "@babel/preset-react": "^7.10.4",
     "@babel/preset-typescript": "^7.10.4",
     "@react-native-community/bob": "^0.16.2",
     "@react-native-community/eslint-config": "2.0.0",
@@ -138,14 +136,12 @@
     "@types/styled-components": "^5.1.3",
     "@typescript-eslint/eslint-plugin": "^4.3.0",
     "@typescript-eslint/parser": "^4.3.0",
-    "ast-pretty-print": "^2.0.1",
     "babel-eslint": "^10.1.0",
     "babel-jest": "^26.3.0",
     "babel-loader": "^8.1.0",
     "babel-plugin-i18next-extract": "^0.7.2",
     "babel-plugin-module-resolver": "^4.0.0",
     "babel-plugin-react-native-web": "^0.13.13",
-    "babel-preset-expo": "^5.0.0",
     "eslint": "^7.10.0",
     "eslint-config-prettier": "^6.12.0",
     "eslint-plugin-babel": "^5.3.1",
