+++ src/index.ts
@@ -3,7 +3,6 @@ export { registerNativeHandlers, NetInfo } from './native';
 export * from './contexts';
 
 export * from './styles/styledComponents';
-export * from './styles/theme';
 export * from './styles/themeConstants';
 
 export * from './utils/Streami18n';
